src/main/java/com/backEnd/clinicaOdontologica/service/PacienteService.java
package com.backEnd.clinicaOdontologica.service;


import com.backEnd.clinicaOdontologica.Repository.PacienteRepository;
import com.backEnd.clinicaOdontologica.entity.Paciente;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class PacienteService {
    @Autowired
    private PacienteRepository pacienteRepository;

    //METODOS
    public Paciente guardarPaciente(Paciente paciente){
        return pacienteRepository.save(paciente);
    }
    public void actualizarPaciente(Paciente paciente){
        pacienteRepository.save(paciente);
    }
    public void eliminarPaciente(Long id){
        pacienteRepository.deleteById(id);
    }
    public Optional<Paciente> buscarPorId(Long id){
        return pacienteRepository.findById(id);
    }
    public Optional<Paciente> buscarPorEmail(String email){
        return pacienteRepository.findByEmail(email);
    }
    public List<Paciente> obtenerTodos(){
        return pacienteRepository.findAll();
    }
}
